% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_gpx.R
\name{parse_gpx}
\alias{parse_gpx}
\title{Parse GPX data file to a data frame}
\usage{
parse_gpx(NAME = "filename", writeData = FALSE, timeZone = "GMT")
}
\arguments{
\item{NAME}{The input argument NAME is the string path to the selected GPX file (gpx format).}

\item{writeData}{Write to disk flag. Default setting is FALSE. If set to TRUE, the parsed GPX data frame is saved as csv file to the same directory as the source GPX file.}

\item{timeZone}{Time zone definition of the parsed data. Default value is "GMT". For other possible values see \code{\link[base]{timezones}}.}
}
\value{
The function returns a data frame with GPS data. If the writeData flag is set to TRUE, the data frame is also saved in csv format to the same location as the source gpx file.
}
\description{
This script reads a gpx file, converts date-time string data. The final gps dataframe is loaded to environment and can be saved as a csv file.
}
\examples{
data_file <- system.file("extdata", "myGPXData.gpx", package="parseGPX")
gpx <- parse_gpx(data_file)
gpx <- parse_gpx(data_file, writeData = TRUE)
gpx <- parse_gpx(data_file, timeZone = "Europe/Ljubljana")
gpx <- parse_gpx(data_file, writeData = TRUE, timeZone = "Europe/Ljubljana")
}
\keyword{GPX}
\keyword{csv}
\keyword{data,}
\keyword{format}
\keyword{parse,}

